move logout to dashboard drawer
