+++ src/components/Drawer.tsx
@@ -6,9 +6,14 @@ import {
   RocketIcon,
 } from "@radix-ui/react-icons";
 import * as NavigationMenu from "@radix-ui/react-navigation-menu";
+import { Icon } from "@radix-ui/react-select";
 import clsx from "clsx";
 import NextLink from "next/link";
 import { usePathname } from "next/navigation";
+
+import { User } from "~/lib/config";
+import fetchJson from "~/lib/fetchJson";
+import useUser from "~/lib/useUser";
 
 const MenuItem = ({ href, Icon, text, ...props }) => {
   const pathname = usePathname();
@@ -27,6 +32,15 @@ const MenuItem = ({ href, Icon, text, ...props }) => {
 };
 
 function Drawer() {
+  const { mutate: mutateUser } = useUser();
+
+  const signOut = async () => {
+    await fetchJson<User>("/api/logout", {
+      method: "POST",
+    });
+    mutateUser(undefined);
+  };
+
   return (
     <NavigationMenu.Root className="mr-2 shrink-0 self-stretch bg-white p-2">
       <NavigationMenu.List>
@@ -47,6 +61,14 @@ function Drawer() {
           Icon={BarChartIcon}
           text="History"
         />
+
+        <NavigationMenu.Item
+          className="rounded px-4 py-2 hover:bg-gray-200"
+          onClick={() => signOut()}
+        >
+          <Icon className="mr-2 inline h-4 w-4 align-middle" />
+          Logout
+        </NavigationMenu.Item>
       </NavigationMenu.List>
     </NavigationMenu.Root>
   );
